Refactor ledger-tool moving some of functions to a separate module
